refactor(Counter): rmove Counter component
